+++ config/initializers/action_view.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-ActiveSupport.on_load :action_view do
+ActiveSupport.on_load :action_view do # rubocop:disable Metrics/BlockLength
 	concerning :DecoratedAssignments, prepend: true do
 		def assign(assignments, ...)
 			decorate assignments
@@ -19,6 +19,23 @@ ActiveSupport.on_load :action_view do
 		def decorate objects
 			objects
 					.transform_values!(&:decorated)
+		end
+	end
+
+	concerning :PresenterContext, prepend: true do
+		def in_rendering_context(...)
+			Magic::Presenter::Base.with view_context: self do
+				super
+			end
+		end
+
+		private
+
+		def decorate(...)
+			super
+					.each_value
+					.grep(Magic::Presenter::Base)
+					.each { _1.view_context = self }
 		end
 	end
 end